process_io: use poll, detect socket close
This fixes the case where a socket service exits before we send
EOF to it. In the case of normal services, we detected SIGCHLD,
but here we need to exit when the socket is closed.

Use poll() for that, as select() will only wake us for the readable
FDs (stdout, stderr), not writable (stdin).

Make sure that also half-close is handled correctly, both ways.